Fix tests build: scope CmuxSettings import, refresh file length budgets
Importing CmuxSettings into WorkspaceUnitTests made StoredShortcut (and
other names) ambiguous against the app target's types across existing
tests. Move the package/app action parity test into
KeyboardShortcutContextTests, which already imports both modules, and
drop the broad import. Refresh swift-file-length-budget entries for the
files this PR grows.
