+++ .github/swift-file-length-budget.tsv
@@ -3,16 +3,16 @@
 # Reduce counts as files shrink. CI fails if tracked files exceed this budget.
 32966	CLI/cmux.swift
 22840	Sources/TerminalController.swift
-19955	Sources/Workspace.swift
-19256	Sources/ContentView.swift
-18057	Sources/AppDelegate.swift
+19985	Sources/Workspace.swift
+19275	Sources/ContentView.swift
+18120	Sources/AppDelegate.swift
 16539	Sources/GhosttyTerminalView.swift
 13608	Sources/Panels/BrowserPanel.swift
-11916	cmuxTests/AppDelegateShortcutRoutingTests.swift
-9992	Sources/TabManager.swift
+12044	cmuxTests/AppDelegateShortcutRoutingTests.swift
+10020	Sources/TabManager.swift
 8494	cmuxTests/CLINotifyProcessIntegrationRegressionTests.swift
 7737	Sources/Panels/BrowserPanelView.swift
-7198	cmuxTests/WorkspaceUnitTests.swift
+7291	cmuxTests/WorkspaceUnitTests.swift
 6948	cmuxTests/WorkspaceRemoteConnectionTests.swift
 6542	cmuxTests/GhosttyConfigTests.swift
 6299	cmuxTests/TerminalAndGhosttyTests.swift
@@ -20,7 +20,7 @@
 6153	CLI/cmux_open.swift
 6071	Sources/TextBoxInput.swift
 5482	cmuxTests/BrowserConfigTests.swift
-5448	Sources/cmuxApp.swift
+5459	Sources/cmuxApp.swift
 4460	Sources/Panels/FilePreviewPanel.swift
 4400	cmuxTests/BrowserPanelTests.swift
 4288	Packages/CmuxMobileShell/Sources/CmuxMobileShell/MobileShellComposite.swift
@@ -38,8 +38,8 @@
 2654	Sources/Panels/CmuxWebView.swift
 2545	cmuxTests/WorkspaceManualUnreadTests.swift
 2527	Sources/TerminalNotificationStore.swift
-2513	cmuxTests/CommandPaletteSearchEngineTests.swift
-2509	Sources/KeyboardShortcutSettings.swift
+2544	cmuxTests/CommandPaletteSearchEngineTests.swift
+2516	Sources/KeyboardShortcutSettings.swift
 2327	cmuxTests/CJKIMEInputTests.swift
 2290	Sources/FileExplorerView.swift
 2260	Sources/TerminalWindowPortal.swift
@@ -129,7 +129,7 @@
 680	Sources/FileExplorerSearchController.swift
 683	Packages/CmuxSwiftRender/Sources/CmuxSwiftRender/SwiftViewInterpreter.swift
 668	cmuxTests/FeedCoordinatorTests.swift
-668	cmuxTests/KeyboardShortcutContextTests.swift
+679	cmuxTests/KeyboardShortcutContextTests.swift
 654	Packages/CmuxSettingsUI/Sources/CmuxSettingsUI/Sections/KeyboardShortcutsSection.swift
 650	Sources/Panels/MarkdownRemoteImageLoader.swift
 649	Sources/CmuxTopSnapshot.swift
